tune search breadth bottlenecks

+++ Search.h
@@ -3,7 +3,7 @@
 
 #include "Board.h"
 
-float alpha_beta(Board* B, int nb_plies, float alpha, float beta);
+int alpha_beta(Board* B, int nb_plies, int alpha, int beta);
 void order_moves(Board* B, MoveList* ML);
 Move get_best_move(Board* B, int nb_plies);
